Create genetic groups regarding countries

Overall function to create genetic groups regarding countries

create_GG(
  psInputFile,
  psBreedsFile = NULL,
  psCountryFile = NULL,
  pminGroupSize = 250,
  pnumberOfYears = 6,
  pglobal_minyear = 1900,
  pglobal_maxyear = 2100,
  pdiffYear = 3,
  pshiftYear = 2,
  psOutputFile,
  pbLog = FALSE
)

Arguments

psInputFile
file including number of missing parents per year and breed

psBreedsFile
Read a breed grouping file

psCountryFile
country grouping file

pminGroupSize
minimum number of animals required per genetic group

pnumberOfYears
minimum number of birth years required per genetic group

pglobal_minyear
global minimum year

pglobal_maxyear
globar maximum year

pdiffYear
difference in birth years between parents of bulls and parents of dams

pshiftYear
shift between genetic groups of sire and genetic groups of dams; this is required so that they do not contain exactly the same animals

psOutputFile
genetic group csv-file